refactor: info 구조체, init_info 함수 분리 및 새로운 디렉토리 생성
